requirements.txt
agno==1.2.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
baidusearch==1.0.3
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Crawl4AI==0.5.0.post4
cryptography==44.0.2
cssselect==1.3.0
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==7.5.3
fake-http-header==0.3.5
fake-useragent==2.1.0
faust-cchardet==2.1.19
feedparser==6.0.11
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanize==4.12.2
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.63.14
lxml==5.3.1
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.2.0
narwhals==1.32.0
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==2.2.4
openai==1.68.2
packaging==24.2
pandas==2.2.3
pillow==10.4.0
playwright==1.51.0
primp==0.14.0
propcache==0.3.0
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pyee==12.1.1
Pygments==2.19.1
pyOpenSSL==25.0.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
resend==2.6.0
rich==13.9.4
rpds-py==0.23.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
streamlit==1.43.2
tenacity==9.0.0
tf-playwright-stealth==1.1.2
tiktoken==0.9.0
tldextract==5.1.3
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0